pyproject.toml
[tool.poetry]
name = "pact-testing-demo"
version = "0.1.0"
description = "Demo projet for PACT testing"
authors = ["fmarayeu"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.1.0"
requests = "^2.32.3"
pytest = "^8.3.3"
pact-python = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
